+++ apps/web/app/(navigation)/budgets/BudgetsTranslated.tsx
@@ -4,7 +4,7 @@ import { useTranslate } from "lib/hooks";
 import { DummyAsideCardContent } from "app/DummyAsideCardContent";
 import MultiCardLayout from "../MultiCardLayout";
 
-import { TransactionsTable } from "../../../../../packages/ui/TransactionsTable";
+import { TransactionsTable } from "./[id]/TransactionsTable";
 
 const budget = {
   id: "71ee9a04-6b27-423a-9caa-7d6a92335dae",
@@ -23,7 +23,10 @@ const budget = {
 export const BudgetsTitle = ({ name = "" }) => {
   const { t, dict } = useTranslate("BudgetsPage");
   const mainCardContent = (
-    <TransactionsTable budget={budget} setSorting={column => console.log(column)} />
+    <TransactionsTable
+      budget={budget}
+      setSorting={(column) => console.log(column)}
+    />
   );
 
   return (
